// src/models/filetype.enum.ts
export enum FileType {
    CLS = 0,
    EXAMPLE = 1,
    GO = 2,
    JAVA = 3,
    JAVASCRIPT = 4,
    PHP = 5,
    OBJECTIVE_C = 6,
    PYTHON = 7,
    SWIFT = 8,
    TERRAFORM = 9,
    YAML = 10,
    C_SHARP = 11,
    C = 12,
    C_PLUS_PLUS = 13,
    CONFIG = 14,
    INI = 15,
    PROPERTIES = 16,
    TOML = 17,
    OTHER = 18,
}